fix(site): keep the browser engine's graph node-free; retire LS in the type test
collectVocabularyNotes lived in scan.ts and scan-files.ts imported it as a
VALUE, which pulled the node-only graph into the browser bundle:

  dist/scan-files.js → dist/scan.js → core/mcp.js → core/refs.js
    → core/symbols.js → @ast-grep/napi (a native .node binding esbuild
      has no loader for)

Measured by walking the built CJS require graph, and confirmed by mutation:
restoring the scan.js import makes the chain above reappear, removing it makes
@ast-grep/napi unreachable. The functions are pure data transforms and now sit
in scan-core.ts, the module whose stated job is to be node-free.

test/types/purity.ts still declared LS, which left the catalog on 2026-08-17.
It is now a negative case instead: a retired name has to FAIL at tsc time,
since its absence from a list proves nothing.
